Added file with LDAP class
