+++ cornflow/shared/authentication.py
@@ -2,7 +2,6 @@ import datetime
 from flask import request, g, current_app
 from functools import wraps
 import jwt
-from ldap3 import Server, Connection, ALL
 
 from ..models.user import UserModel
 from ..shared.exceptions import InvalidCredentials, ObjectDoesNotExist, NoPermission
@@ -123,56 +122,3 @@ class Auth:
     def return_user_from_token(token):
         user_id = Auth.decode_token(token)["user_id"]
         return user_id
-
-    @staticmethod
-    def get_bound_connection():
-        if "ldpa_connection" in g:
-            return g.ldap_connection
-        server = Server(current_app.config["LDAP_HOST"], get_info=ALL)
-        g.ldap_connection = Connection(
-            server,
-            current_app.config["LDAP_BIND_DN"],
-            current_app.config["LDAP_BIND_PASSWORD"],
-            auto_bind=True,
-        )
-        return g.ldap_connection
-
-    @staticmethod
-    def get_dn_from_user(user):
-        return "%s=%s,%s" % (
-            current_app.config["LDAP_USERNAME_ATTRIBUTE"],
-            user,
-            current_app.config["LDAP_USER_BASE"],
-        )
-
-    @staticmethod
-    def get_user_email(user):
-        email_attribute = current_app.config.get("LDAP_EMAIL_ATTRIBUTE", False)
-        if not email_attribute:
-            return False
-        conn = Auth.get_bound_connection()
-        user_search = Auth.get_dn_from_user(user)
-
-        user_object = "(objectclass=%s)" % (
-            current_app.config["LDAP_USER_OBJECT_CLASS"],
-        )
-
-        conn.search(user_search, user_object, attributes=[email_attribute])
-
-        if len(conn.entries) < 1:
-            return False
-
-        return getattr(conn.entries[0], email_attribute, False)[0]
-
-    @staticmethod
-    def ldap_authenticate(user, password):
-        s = Server(current_app.config["LDAP_HOST"], get_info=ALL)
-        user_dn = Auth.get_dn_from_user(user)
-
-        c = Connection(current_app.config["LDAP_HOST"], user=user_dn, password=password)
-
-        if not c.bind():
-            print("Unable to bind user")
-            return False
-
-        return True
